feat: add tests

+++ src/useCases/TestBenches/DeleteTestBench/DeleteTestBenchUseCase.ts
@@ -1,4 +1,5 @@
 import { ITestBenchRepository } from "src/repositories/TestBench/ITestBenchRepository";
+import { TestBenchNotFound } from "src/utils/errors/TestBenchError";
 
 import { IDeleteTestBenchRequestDTO } from "./DeleteTestBenchDTO";
 
@@ -7,6 +8,10 @@ export class DeleteTestBenchUseCase {
 
 	async execute(data: IDeleteTestBenchRequestDTO) {
 		const testBench = await this.testBenchRepository.findById(data.testBenchId);
+
+		if (!testBench) {
+			throw new TestBenchNotFound();
+		}
 
 		await this.testBenchRepository.delete(testBench.id);
 	}
